syntax_ext: Reuse built-in attribute template checking for macro attributes
