+++ src/test/ui/allocator/allocator-args.stderr
@@ -2,7 +2,7 @@ error: malformed `global_allocator` attribute input
   --> $DIR/allocator-args.rs:10:1
    |
 LL | #[global_allocator(malloc)]
-   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: must be of the form: ``#[global_allocator]``
+   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: must be of the form: `#[global_allocator]`
 
 error: aborting due to previous error
